♻️ refactor: Integrate reducer tests
